fix: moved core Tree logic to BaseTree to avoid circular dependency
